requirements.txt
mkdocs>=1.5.0
mkdocs-material>=9.0.0
mkdocs-exporter>=6.0.0
playwright>=1.40.0
pyyaml>=6.0.0
